project/requirements.txt
streamlit>=1.24.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
geopy>=2.3.0
pandas>=2.0.0
numpy>=1.24.0 